Move `silly-wisher-names` script into executor
