+++ libs/silly-wisher-names/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@genshin-optimizer/silly-wisher-names",
   "version": "0.0.1",
-  "type": "commonjs"
+  "type": "commonjs",
+  "executors": "./executors.json"
 }
