+++ libs/silly-wisher-names/src/executors/gen-assets/schema.d.ts
@@ -1,0 +1,1 @@
+export interface GenAssetsExecutorSchema {} // eslint-disable-line
