+++ libs/silly-wisher-names/src/executors/gen-assets/schema.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "version": 2,
+  "title": "GenAssets executor",
+  "description": "",
+  "type": "object",
+  "properties": {},
+  "required": []
+}
